ci(GHA): Resolve commitlint failures
